ComponentScan Test example
Using an inner class, which is a fully qualified java class, can use @ComponentScan to scan the components. Because Laurel is annotated with @Primary, it is picked up, and is the expected outcome. Note, nothing else is done in TestConfig in this example, except specify it as a Configuration and to set the ComponentScan.