Profiles in tests
Set up a new test to have an ActiveProfile. Needed to modify the class under test to define the Profile and be Primary. Also, ran into problems with componentScan picking up test files, so needed to set @ActiveProfiles in the test classes and @Profile in the config classes so they are profile-specific.